+++ atomica/system.py
@@ -3,37 +3,42 @@ Define internal system constants and functions
 
 """
 
-import os
+import sciris as sc
 from .function_parser import supported_functions
+import pathlib
 
 # Set up a logger that can be imported elsewhere
 import logging
+
 logger = logging.getLogger('atomica')
 
 
-def atomica_path(subdir=None, trailingsep=True):
-    """ Returns paths relative to the Atomica parent module
+def atomica_path(subdir=None) -> pathlib.Path:
+    """
+    Returns paths relative to the Atomica parent module
 
     This function by default returns the directory containing the Atomica
     source files. It can also return paths relative to this directory using
     the optional additional arguments
 
-    :param subdir: Append an additional path list to Atomica path
-    :param trailingsep: If True, a trailing separator will be included so that the
-                        returned path can have a file name string added to it easily
-    :return: Absolute path string
+    Example usage:
+
+        >>> at.atomica_path()
+        WindowsPath('E:/projects/atomica/atomica')
+        >>> at.atomica_path('foo')
+        WindowsPath('E:/projects/atomica/atomica/foo')
+        >>> at.atomica_path(['foo','bar'])
+        WindowsPath('E:/projects/atomica/atomica/foo/bar')
+
+    :param subdir: Optionally append a list of subdirectories to the path
+    :return: Absolute Path object
 
     """
 
-    path = os.path.abspath(os.path.join(os.path.dirname(__file__), os.pardir))
-    if subdir is not None:
-        if not isinstance(subdir, list):
-            subdir = [subdir]  # Ensure it's a list.
-        tojoin = [path] + subdir
-        if trailingsep:
-            tojoin.append('')  # This ensures it ends with a separator.
-        path = os.path.join(*tojoin)  # For example: ['/home/atomica', 'tests', '']
-    return path
+    path = pathlib.Path(__file__).parent
+    for d in sc.promotetolist(subdir):
+        path /= d
+    return path.resolve()
 
 
 LIBRARY_PATH = atomica_path(['atomica', 'library'])