+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This file records changes to the codebase grouped by version release. Unreleased
 
 ## [Unreleased]
 
+- `at.atomica_path` returns a `Path` object. The `trailingsep` argument has been removed as it is not relevant when returning a path 
+- `at.parent_dir` returns a `Path` object
+- `Project.load_databook` now supports passing in a `ProjectData` instance (as does the `Project` constructor) 
+- Remove unused `num_interpops` argument from `Project.create_databook()`
 - Add support for '>' in transition matrix to represent junction residual
 
 ## [1.19.0] - 2020-06-26